+++ r-appservice.tf
@@ -114,3 +114,27 @@ resource "azurerm_app_service" "app_service" {
     ]
   }
 }
+
+resource "azurerm_app_service_certificate" "app_service_certificate" {
+  for_each = var.custom_domains != null ? {
+    for k, v in var.custom_domains :
+    k => v if v != null
+  } : {}
+
+  name                = each.key
+  resource_group_name = var.resource_group_name
+  location            = var.location
+  pfx_blob            = contains(keys(each.value), "certificate_file") ? filebase64(each.value.certificate_file) : null
+  password            = contains(keys(each.value), "certificate_file") ? each.value.certificate_password : null
+  key_vault_secret_id = contains(keys(each.value), "certificate_keyvault_certificate_id") ? each.value.certificate_keyvault_certificate_id : null
+}
+
+resource "azurerm_app_service_custom_hostname_binding" "app_service_custom_hostname_binding" {
+  for_each = var.custom_domains != null ? var.custom_domains : {}
+
+  hostname            = each.key
+  app_service_name    = azurerm_app_service.app_service.name
+  resource_group_name = var.resource_group_name
+  ssl_state           = lookup(azurerm_app_service_certificate.app_service_certificate, each.key, false) != false ? "SniEnabled" : null
+  thumbprint          = lookup(azurerm_app_service_certificate.app_service_certificate, each.key, false) != false ? azurerm_app_service_certificate.app_service_certificate[each.key].thumbprint : null
+}